+++ spec/acceptance/suites/default/install_simp_deps_repo_spec.rb
@@ -2,69 +2,40 @@ require 'spec_helper_acceptance'
 
 hosts.each do |host|
   describe '#write_hieradata_to' do
-    let (:default_repo_content)  { <<-EOF
-# SIMP Dependencies Yum Repository
-[simp_dependencies]
-name=simp-project-6_X_Dependencies
-gpgcheck=1
-enabled=1
-baseurl=https://packagecloud.io/simp-project/6_X_Dependencies/el/$releasever/$basearch
-gpgkey=https://raw.githubusercontent.com/NationalSecurityAgency/SIMP/master/GPGKEYS/RPM-GPG-KEY-SIMP
-       https://yum.puppet.com/RPM-GPG-KEY-puppetlabs
-       https://yum.puppet.com/RPM-GPG-KEY-puppet
-       https://apt.postgresql.org/pub/repos/yum/RPM-GPG-KEY-PGDG-96
-       https://artifacts.elastic.co/GPG-KEY-elasticsearch
-       https://grafanarel.s3.amazonaws.com/RPM-GPG-KEY-grafana
-       https://dl.fedoraproject.org/pub/epel/RPM-GPG-KEY-EPEL-$releasever
-sslverify=1
-sslcacert=/etc/pki/tls/certs/ca-bundle.crt
-metadata_expire=300
-EOF
-    }
-    let (:nondefault_repo_content)  { <<-EOF
-# SIMP Dependencies Yum Repository
-[simp_dependencies]
-name=simp-project-7_X_dependencies
-gpgcheck=0
-enabled=1
-baseurl=https://packagecloud.io/simp-project/7_X_dependencies/el/$releasever/$basearch
-gpgkey=https://raw.githubusercontent.com/NationalSecurityAgency/SIMP/master/GPGKEYS/RPM-GPG-KEY-SIMP
-       https://yum.puppet.com/RPM-GPG-KEY-puppetlabs
-       https://yum.puppet.com/RPM-GPG-KEY-puppet
-       https://apt.postgresql.org/pub/repos/yum/RPM-GPG-KEY-PGDG-96
-       https://artifacts.elastic.co/GPG-KEY-elasticsearch
-       https://grafanarel.s3.amazonaws.com/RPM-GPG-KEY-grafana
-       https://dl.fedoraproject.org/pub/epel/RPM-GPG-KEY-EPEL-$releasever
-sslverify=1
-sslcacert=/etc/pki/tls/certs/ca-bundle.crt
-metadata_expire=300
-EOF
-    }
+
+    it 'should install yum utils' do
+      host.install_package('yum-utils')
+    end
+
     context 'defailt settings' do
-      before(:all) { install_simp_deps_repo(host) }
-      after(:all) { on(host, "rm -rf /etc/yum.repos.d/simp_dependencies.repo") }
+      before(:all) { install_simp_repos(host) }
 
       it 'creates the repo' do
-        on host, 'test -f /etc/yum.repos.d/simp_dependencies.repo'
+        on host, 'test -f /etc/yum.repos.d/simp6.repo'
+        on host, 'test -f /etc/yum.repos.d/simp6_deps.repo'
       end
 
-      it 'writes the correct contents to the correct file' do
-        stdout = on(host, 'cat /etc/yum.repos.d/simp_dependencies.repo').stdout
-        expect(stdout).to eq(default_repo_content)
+      it 'enables the correct repos' do
+        simp6info = on(host, '/usr/bin/yum repolist -v simp6 | grep ^Repo-status').stdout.strip
+        expect(simp6info).to match(/.*Repo-status.*enabled.*/)
+        simp6depsinfo = on(host, 'yum repolist -v simp6_deps| grep ^Repo-status').stdout.strip
+        expect(simp6depsinfo).to match(/.*Repo-status.*enabled.*/)
       end
     end
 
-    context 'when passed a hash' do
-      before(:all) { install_simp_deps_repo(host, '7_X_dependencies', '0' ) }
-      after(:all) { on(host, "rm -rf /etc/yum.repos.d/simp_dependencies.repo") }
+    context 'when passed a disabled list ' do
+      before(:all) { install_simp_repos(host, ['simp6'] ) }
 
       it 'creates the repo' do
-        on host, 'test -f /etc/yum.repos.d/simp_dependencies.repo'
+        on host, 'test -f /etc/yum.repos.d/simp6.repo'
+        on host, 'test -f /etc/yum.repos.d/simp6_deps.repo'
       end
 
-      it 'writes the correct contents to the correct file' do
-        stdout = on(host, 'cat /etc/yum.repos.d/simp_dependencies.repo').stdout
-        expect(stdout).to eq(nondefault_repo_content)
+      it 'enables the correct repos' do
+        simp6info = on(host, 'yum repolist -v simp6 | grep ^Repo-status').stdout.strip
+        expect(simp6info).to match(/.*Repo-status.*disabled.*/)
+        simp6depsinfo = on(host, 'yum repolist -v simp6_deps| grep ^Repo-status').stdout.strip
+        expect(simp6depsinfo).to match(/.*Repo-status.*enabled.*/)
       end
     end
   end